requirements.txt
aiogram==3.17.0
yt-dlp
python-dotenv
pdf2docx
python-docx
openpyxl
python-pptx
pymupdf
shazamio
comtypes
docx2pdf
wikipedia
g4f
gTTS
deep-translator
